nothing of note

+++ pytorch/words/train.py
@@ -36,7 +36,7 @@ def median(str_list):
 
 
 def subi(sub:str): # substring to index 
-    return torch.long.from_bytes(sub.encode('utf-8'), 'big')
+    return int.from_bytes(sub.encode('utf-8'), 'big')
 
 def isub(i): # index to substring
     return i.to_bytes((i.bit_length() + 7) // 8, 'big').decode('utf-8')
